chore(ZMSKVR-795 GH-1342): refactor calendarview for better maintainability (#1499)
* chore(ZMSKVR-795 GH-1342): extract ProviderView and refactor AppointmentSelectionView to delegate provider UI

- Move provider checkbox/teaser into new ProviderView
- Keep API fetching and state in AppointmentSelectionView
- Fix recursive updates by removing v-model mirroring and emitting explicit toggles
- Update imports and wiring; adjust tests

* chore(ZMSKVR-795 GH-1342): refactor AppointmentSelectionView into smaller components

- Split large AppointmentSelectionView into 6 focused components:
  - ProviderView: handles provider selection (multi-provider checkboxes + single provider teaser)
  - CalendarView: handles calendar-based appointment selection with hourly/day-part views
  - ListView: handles accordion-style list view with available days
  - ViewToggle: handles calendar/list view switching with internal state management
  - AppointmentLayout: shared component for appointment time slot display
  - SelectedAppointmentSummary: displays selected appointment details

- Organized components in AppointmentSelection/ folder for better structure
- Moved toggle logic and styles to ViewToggle component for better encapsulation
- Used AppointmentLayout for consistent styling across CalendarView and ListView
- Maintained all existing functionality while improving maintainability
- Fixed recursive update issues with proper event handling and state management

* chore(ZMSKVR-795 GH-1342): extract provider helpers, centralize formatting and constants\n\n- Add getOfficeById/officeNameById and update usages in children\n- Use formatAppointmentDateTime util for day/time formatting\n- Move APPOINTMENTS_THRESHOLD_FOR_HOURLY_VIEW to utils/Constants and import in children\n- Pass openAccordionDate down; ListView emits toggleDay to parent\n- Minor prop/type cleanups in CalendarView/ListView

* chore(ZMSKVR-795 GH-1342): factor view-specific logic into subcomponents\n\n- ListView: own firstFiveAvailableDays, load more, earlier/later, snap, accordion\n- ListView: compute day availability helpers locally; expose snap via defineExpose\n- CalendarView: move snap, earlier/later, available-hours; expose snap; emit setters\n- Parent: wire refs to child snaps; remove duplicate/list-specific logic\n- SelectedAppointmentSummary: compute estimated duration locally\n- Minor prop/type cleanups and event wiring updates

* fix(ZMSKVR-795 GH-1342): fix a bunch of zmscitizienview tests from refactor

* chore(ZMSKVR-795 GH-1342): split AppointmentSelection tests into focused component specs and reorganize describes

- Move provider selection tests to tests/unit/Appointment/AppointmentSelection/ProviderSelection.spec.ts

- Add CalendarListToggle.spec.ts and move toggle test

- Add AppointmentPreview.spec.ts with focused unit tests

- Add ListView.spec.ts and CalendarView.spec.ts; migrate list/calendar tests

- Create TimeSlotGrid.spec.ts and move slot-grid related tests

- Rename components: AppointmentSelectionView->AppointmentSelection, SelectedAppointmentSummary->AppointmentPreview, ViewToggle->CalendarListToggle, ProviderView->ProviderSelection, AppointmentLayout->TimeSlotGrid

- Update CalendarView.vue and ListView.vue; keep integration tests in AppointmentSelection.spec.ts minimal

- Reorganize AppointmentSelection.spec.ts into describe blocks (ProviderSelection Integration, CalendarView Integration, CalendarListToggle Integration, Error States Integration, Submission/Loading State Integration)

* fix(ZMSKVR-795 GH-1342): wrong name appointment-selection-view is now AppointmentSelection

* fix(ZMSKVR-795 GH-1342): Fix invalid CSS value justify-content: left is invalid; use flex-start.

* fix(ZMSKVR-795 GH-1342): Fix condition: callout renders when timeslot is undefined. selectedTimeslot !== 0 is true for undefined, so the callout can show without a valid time. Guard for a positive numeric timeslot.

* fix(ZMSKVR-795 GH-1342): Bug: midnight (0:00) hours are ignored when initializing after fetch.

* fix(ZMSKVR-795 GH-1342): Preserve preselected/initial office as the first shown and selected. Sorting reorders offices and may override the intended first office. Capture it before sorting and pass it to showSelectionForProvider. Bug: midnight (0:00) hours are ignored when initializing after fetch.

* fix(ZMSKVR-795 GH-1342): Guard against missing clean/restricted providers to avoid runtime errors. find() may return undefined; current code asserts non-null and can crash.

* fix(ZMSKVR-795 GH-1342): Use ASCII hyphen in hour label to match tests and avoid copy/paste chars.

* fix(ZMSKVR-795 GH-1342): Fix aria-controls value (remove '#').

* fix(ZMSKVR-795 GH-1342): Clean up style of earlier and later buttons in CalendarView and ListView

* fix(ZMSKVR-795 GH-1342): Fix canLoadMore logic to use truly available days. Currently compares visible slice to raw availableDays length, showing “Load more” even when no more truly-available days exist.


* fix(ZMSKVR-795 GH-1342): Loading state never completes: isLoadingComplete stays false. isLoadingComplete is set to false when fetching starts but never set to true, so consumers (e.g., CalendarView) may think loading never finishes.

* fix(ZMSKVR-795 GH-1342): Map mutations are not reactive; UI can fail to update. ref(new Map()) + .set() won't trigger updates. Reassign a new Map in setters (already present), and use those setters everywhere mutations occur.

* fix(ZMSKVR-795 GH-1342): Bug: time labels hidden when earliest hour is 0. The label depends on firstHour > 0, so midnight (0) suppresses all labels.

* fix(ZMSKVR-795 GH-1342): Fix: summary ref points to a component, not an HTMLElement (runtime crash on focus/scroll).Calling focus()/scrollIntoView() on a component ref throws at runtime unless the child exposes DOM methods. Wrap the child in a DOM element or expose a focus() method.

* fix(ZMSKVR-795 GH-1342): Make the toggle controlled via v-model and add an accessible name. Currently the component manages its own isListView state, which desynchronizes it from the parent. Expose isListView as a prop and only emit updates. Also, role=switch needs an accessible name; bind aria-labelledby to the H2.

* fix(ZMSKVR-795 GH-1342): stabilize nearest-date snapping test after provider change

- Control subsequent fetchAvailableDays responses during provider-change pipeline to return only 2025-06-17 on the final refetch\n- Trigger deep watcher by replacing selectedProviders object instead of mutating\n- Add sufficient awaits to cover debounce + async fetch before asserting selectedDay\n- Keep UI logic unchanged; tests now reflect intended UX behavior

* fix(ZMSKVR-795 GH-1342): Padding issues

* docs(ZMSKVR-110): document rationale for APPOINTMENTS_THRESHOLD_FOR_HOURLY_VIEW=18

- Reference ZMSKVR-110: few appointments (≤18) use am/pm; many (>18) use hourly\n- Adds short comment to prevent hard-coded drift and tie to UX rationale

* Cleanup zmskvr 511 fix date padding in listview (#1533)

* clean(ZMSKVR): Refactor some components to a better place in zmscitizenview

* clean(ZMSKVR): fix some lint errors in CustomerInfo.vue

---------
