+++ zmscitizenview/tests/unit/Appointment/ServiceFinder/SubserviceListItem.spec.ts
@@ -3,7 +3,7 @@ import { describe, expect, it, beforeEach } from "vitest";
 import { nextTick } from "vue";
 
 // @ts-expect-error: Vue SFC import for test
-import SubserviceListItem from "@/components/Appointment/SubserviceListItem.vue";
+import SubserviceListItem from "@/components/Appointment/ServiceFinder/SubserviceListItem.vue";
 
 describe("SubserviceListItem", () => {
   const mockSubService = {
